Add checking before removing loan rules.

+++ admin/modules/circulation/loan_rules.php
@@ -104,15 +104,22 @@ if (isset($_POST['saveData'])) {
     // loop array
     foreach ($_POST['itemID'] as $itemID) {
         $itemID = (integer)$itemID;
-        if (!$sql_op->delete('mst_loan_rules', 'loan_rules_id='.$itemID)) {
-            $error_num++;
+        $lrStatus = circapi::is_any_active_loanrules($dbs, $itemID);
+        if (!$lrStatus) {
+            if (!$sql_op->delete('mst_loan_rules', 'loan_rules_id='.$itemID)) {
+                $error_num++;
+            }
         }
     }
-
     // error alerting
     if ($error_num == 0) {
-        utility::jsAlert(__('All Data Successfully Deleted'));
-        echo '<script language="Javascript">parent.jQuery(\'#mainContent\').simbioAJAX(\''.$_SERVER['PHP_SELF'].'?'.$_POST['lastQueryStr'].'\');</script>';
+        if (!$lrStatus) {
+            utility::jsAlert(__('All Data Successfully Deleted'));
+            echo '<script language="Javascript">parent.jQuery(\'#mainContent\').simbioAJAX(\''.$_SERVER['PHP_SELF'].'?'.$_POST['lastQueryStr'].'\');</script>';
+        } else {
+            utility::jsAlert(__('Sorry. There is active loan transaction(s) using this loan rules.'));
+            echo '<script language="Javascript">parent.jQuery(\'#mainContent\').simbioAJAX(\''.$_SERVER['PHP_SELF'].'\');</script>';
+        }
     } else {
         utility::jsAlert(__('Some or All Data NOT deleted successfully!\nPlease contact system administrator'));
         echo '<script language="Javascript">parent.jQuery(\'#mainContent\').simbioAJAX(\''.$_SERVER['PHP_SELF'].'?'.$_POST['lastQueryStr'].'\');</script>';
